Add exception handling to MovieServiceImpl

+++ movie-recommender-service/src/main/java/cz/muni/fi/pa165/MovieServiceImpl.java
@@ -2,6 +2,7 @@ package cz.muni.fi.pa165;
 
 import cz.muni.fi.pa165.dao.MovieDao;
 import cz.muni.fi.pa165.entity.Movie;
+import cz.muni.fi.pa165.exceptions.ServiceLayerException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -26,31 +27,89 @@ public class MovieServiceImpl implements MovieService {
 
     @Override
     public void create(Movie movie) {
-        movieDao.create(movie);
+        if (movie == null) {
+            throw new IllegalArgumentException("Movie parameter is null.");
+        }
+
+        try {
+            movieDao.create(movie);
+        } catch (Exception e) {
+            throw new ServiceLayerException("Error occurred while creating Movie.", e);
+        }
     }
 
     @Override
     public List<Movie> findAll() {
-        return movieDao.findAll();
+        try {
+            return movieDao.findAll();
+        } catch (Exception e) {
+            throw new ServiceLayerException("Error occurred while retrieving all movies.");
+        }
     }
 
     @Override
-    public Movie findById(long id) {
-        return movieDao.findById(id);
+    public Movie findById(Long id) {
+        if (id == null) {
+            throw new IllegalArgumentException("Id parameter is null.");
+        }
+
+        try {
+            return movieDao.findById(id);
+        } catch (Exception e) {
+            throw new ServiceLayerException("Error occurred while retrieving movie by id.", e);
+        }
     }
 
     @Override
     public List<Movie> findByTitle(String title) {
-        return movieDao.findByTitle(title);
+        if (title == null) {
+            throw new IllegalArgumentException("Title parameter is null.");
+        }
+
+        if (title.isEmpty()) {
+            throw new IllegalArgumentException("Title cannot be empty.");
+        }
+
+        try {
+            return movieDao.findByTitle(title);
+        } catch (Exception e) {
+            throw new ServiceLayerException("Error occurred while retrieving movie by title.", e);
+        }
     }
 
     @Override
     public void update(Movie movie) {
-        movieDao.update(movie);
+        if (movie == null) {
+            throw new IllegalArgumentException("Movie parameter is null");
+        }
+
+        Movie found = movieDao.findById(movie.getId());
+        if (found == null) {
+            throw new IllegalArgumentException("Movie is not in the database.");
+        }
+
+        try {
+            movieDao.update(movie);
+        } catch (Exception e) {
+            throw new ServiceLayerException("Error occurred while updating movie", e);
+        }
     }
 
     @Override
     public void remove(Movie movie) {
-        movieDao.remove(movie);
+        if (movie == null) {
+            throw new IllegalArgumentException("Movie parameter is null");
+        }
+
+        Movie found = movieDao.findById(movie.getId());
+        if (found == null) {
+            throw new IllegalArgumentException("Movie is not in the database.");
+        }
+
+        try {
+            movieDao.remove(movie);
+        } catch (Exception e) {
+            throw new ServiceLayerException("Error occurred while removing movie", e);
+        }
     }
 }
